Fix select_global for non-sharded tables
As far as I can tell, behavior for selecting the global group
a table that isn't sharded is poorly defined, becuase it requires
a mapping id to look up the global group.

For unsharded tables, there are no mapping entries to find the
shard mapping id, so it errored out.  I added some fallthrough
code to simply return the first defined shard mapping id in this
case, which will suffice for simple setups.  I'm not sure how
it's supposed to work for more complicated setups, but Mats
Kindahl's comment here suggests to me that it is undefined,
and may be reworked in the future:
http://blog.ulf-wendel.de/2013/mysql-5-7-fabric-any-good/#comment-34915

Also add more debug