Re-initialize hash_key before appending
Found it! 🔪🐛

When used in a loop (as in mysqlnd_ms_fabric_select_servers), hash keys
after the first would be dutifully appended to the existing hash key -
after the \0 end byte. Thus, any hash key after the first appeared as if
it were the same as the first, which in this case broke multiple servers
per group.